ensure feature flag auditor records derive the proper root_account_id
if a given feature flag of account context is audited, the
root_account_id should be set properly if the account is a root account.
In the olden' days root accounts would have a `nil` root_account_id
signifying that they were a root account. This is no longer the case
as we have a "dummy" root account that is used for the root account,
that has an id of 0.

fixes FOO-4405
flag = none

test plan:
- ensure that the feature flag auditor records have the proper
  root_account_id set when the account is a root account

Change-Id: I68b00f1ea8cbc5848f7af3c313fd5c8c6a10e793
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/344217
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>