Convert backbone overrides to javascript
The goal of this patchset is to remove the DifferentiatedModulesSection
dependency on Backbone.

This will allow it to be used easier in other places in canvas.

Specifically, in the discussion create/edit page where backbone is not
used.

refs LF-1384
flag=differentiated_modules

Test Plan
1. Differentiated assignTo tray should work normally in assignment
create and edit
2. Check
2a module overrides
2b. setting with and without the "everyone" option
2c. Various override states

Change-Id: I4936c627a9b549e463da36cc19b23254746873e4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/343751
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jonathan Guardado <jonathan.guardado@instructure.com>
QA-Review: Jonathan Guardado <jonathan.guardado@instructure.com>
Product-Review: Jason Gillett <jason.gillett@instructure.com>